working out get request issue
