+++ src/redux/sagas/addChicken.saga.js
@@ -1,18 +1,6 @@
 import { put, takeLatest } from 'redux-saga/effects';
 import axios from 'axios';
 
-function* fetchChickens(action) {
-  console.log('hit fetchChickens', action.url);
-  let response = yield axios({
-    method: 'GET',
-    url: action.url
-  });
-  yield put({
-    type: 'SET_CHICKENS',
-    payload: response.data
-  });
-}
-
 function* createChicken(action) {
   console.log('hit createChicken', action.payload); 
     
@@ -26,9 +14,8 @@ function* createChicken(action) {
     });
 }
 
-function* chickenSaga() {
-  yield takeLatest('FETCH_CHICKENS', fetchChickens);
+function* createChickenSaga() {
   yield takeLatest('CREATE_CHICKEN', createChicken);
 }
 
-export default chickenSaga;
+export default createChickenSaga;
